+++ brokerAgenda/Controllers/CustomerController.cs
@@ -94,6 +94,7 @@ namespace brokerAgenda.Controllers
       {
         return NotFound();
       }
+      ViewBag.isAppointment = _db.Appointments.Any(a => a.IdBroker == id);
       return View(customer);
     }
     
@@ -107,6 +108,9 @@ namespace brokerAgenda.Controllers
       {
         return NotFound();
       }
+      // Check for appointments to prevent deletion if any
+      if (_db.Appointments.Any(a => a.IdCustomer == id)) return RedirectToAction("Delete");
+
       _db.Customers.Remove(customer);
       _db.SaveChanges();
       TempData["modification"] = "delete";
